# backend/public/api-docs.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Task Tracker API",
    "version": "1.0.0",
    "description": "API untuk manajemen project dan task. Autentikasi menggunakan Laravel Sanctum Personal Access Token."
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "API Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "sanctum": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Masukkan token yang didapat dari endpoint /api/auth/login"
      }
    },
    "schemas": {
      "UserResource": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "name": {
            "type": "string",
            "example": "Admin"
          },
          "email": {
            "type": "string",
            "format": "email",
            "example": "admin@tasktracker.com"
          },
          "is_admin": {
            "type": "boolean",
            "example": true
          }
        }
      },
...
    }
  },
  "paths": {
    "/api/auth/login": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Login pengguna",
        "description": "Autentikasi pengguna dengan email dan password. Mengembalikan Personal Access Token dengan waktu ekspirasi.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "password"],
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email",
                    "example": "admin@tasktracker.com"
                  },
                  "password": {
                    "type": "string",
                    "format": "password",
                    "example": "password"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login berhasil",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Login berhasil."
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "token": {
                          "type": "string",
                          "example": "1|abc123..."
                        },
                        "expires_at": {
                          "type": "string",
                          "format": "date-time",
                          "example": "2025-03-07 12:34:56"
                        },
                        "user": {
                          "$ref": "#/components/schemas/UserResource"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Kredensial tidak valid"
          },
          "422": {
            "description": "Validasi gagal"
          }
        }
      }
    },
    "/api/auth/logout": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Logout pengguna",
        "description": "Mencabut (revoke) token aktif yang sedang digunakan.",
        "security": [{"sanctum": []}],
        "responses": {
          "200": {
            "description": "Logout berhasil"
          },
          "401": {
            "description": "Tidak terautentikasi"
          }
        }
      }
    },
    "/api/auth/me": {
      "get": {
        "tags": ["Authentication"],
        "summary": "Profil pengguna yang sedang login",
        "security": [{"sanctum": []}],
        "responses": {
          "200": {
            "description": "Data pengguna",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "$ref": "#/components/schemas/UserResource"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Tidak terautentikasi"
          }
        }
      }
    },
...
  }
}